Fix style issues shown by rubocop linter - 2.
