+++ .idea/workspace.xml
@@ -4,7 +4,8 @@
     <list default="true" id="98c978b3-02cb-4846-8d54-bf94025b9c65" name="Default Changelist" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/bin/main.rb" beforeDir="false" afterPath="$PROJECT_DIR$/bin/main.rb" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/Game.rb" beforeDir="false" afterPath="$PROJECT_DIR$/lib/Game.rb" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/Game.rb" beforeDir="false" afterPath="$PROJECT_DIR$/lib/game.rb" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/Player.rb" beforeDir="false" afterPath="$PROJECT_DIR$/lib/player.rb" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
