+++ lib/player.rb
@@ -1,3 +1,4 @@
+# This a player class to save properties and capabilities of player
 class Player
   attr_accessor :name
   def initialize(name)
